remove rules stdlib
